eRPC updates 12/2017 2
-- Improved code size of generated code.
-- Generating crc value is optional.
-- Fixed CMSIS Uart driver. Removed dependency on KSDK.
-- List names are based on their types. Names are more deterministic.
-- Service objects are as a default created as global static objects.
-- Added missing doxygen comments.
-- Forbid users use reserved words.
-- Removed outByref for function parameters.
-- Added support for 64bit numbers.
-- Added support of program language specific annotations.
-- Optimized code style of callback functions.
-- Use newer version of clang-format application (5.0).
-- Updated and fixed clang-format config file.
-- Updated Copyrights.
-- Added support of Zephyr RTOS.

+++ test/test_unions/test_unions_server_impl.cpp
@@ -1,10 +1,12 @@
 /*
+ * The Clear BSD License
  * Copyright (c) 2014-2016, Freescale Semiconductor, Inc.
  * Copyright 2016-2017 NXP
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
+ * are permitted (subject to the limitations in the disclaimer below) provided
+ * that the following conditions are met:
  *
  * o Redistributions of source code must retain the above copyright notice, this list
  *   of conditions and the following disclaimer.
@@ -17,6 +19,7 @@
  *   contributors may be used to endorse or promote products derived from this
  *   software without specific prior written permission.
  *
+ * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS LICENSE.
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -211,7 +214,7 @@ foo *sendMyUnion(fruit discriminator, const unionType *unionVariable)
             for (uint32_t i = 1; i <= unionVariable->a.elementsCount; ++i)
             {
                 // If data sent across is incorrect, return 0x55
-                if (i != unionVariable->a.elements[i])
+                if ((int32_t)i != unionVariable->a.elements[i])
                 {
                     newFoo->discriminator = returnVal;
                     newFoo->bing.ret = 0x55;
